requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
python-dotenv==1.0.0
yfinance==0.2.32
matplotlib==3.8.2
pandas==2.1.4
